ci: use revive instead of golint if .revive.toml config is present

+++ ci_checks.sh
@@ -5,11 +5,12 @@
 #    1. gofmt         (http://golang.org/cmd/gofmt/)
 #    2. goimports     (http://godoc.org/golang.org/x/tools/cmd/goimports)
 #    3. golint        (https://github.com/golang/lint)
-#    4. go vet        (http://golang.org/cmd/vet)
-#    5. go test       (https://golang.org/cmd/go/#hdr-Test_packages)
-#    6. race detector (http://blog.golang.org/race-detector)
-#    7. test coverage (http://blog.golang.org/cover)
-#    8. GolangCI-Lint (https://github.com/golangci/golangci-lint)
+#    4. revive        (https://github.com/mgechev/revive)
+#    5. go vet        (http://golang.org/cmd/vet)
+#    6. go test       (https://golang.org/cmd/go/#hdr-Test_packages)
+#    7. race detector (http://blog.golang.org/race-detector)
+#    8. test coverage (http://blog.golang.org/cover, https://github.com/mattn/goveralls)
+#    9. GolangCI-Lint (https://github.com/golangci/golangci-lint)
 
 echo "### gofmt"
 test -z "$(find . -type f -name '*.go' -print0 | xargs -0 gofmt -l -s  | tee /dev/stderr)"
@@ -21,10 +22,17 @@ test -z "$(find . -type f -name '*.go' -print0 | xargs -0 goimports -l  | tee /d
 GOIMPORTS_FAIL=$?
 echo
 
-echo "### golint"
-test -z "$(golint ./... | tee /dev/stderr)"
-GOLINT_FAIL=$?
-echo
+if [ -f .revive.toml ]; then
+	echo "### revive"
+	test -z "$(revive -config .revive.toml ./... | tee /dev/stderr)"
+	REVIVE_FAIL=$?
+	echo
+else
+	echo "### golint"
+	test -z "$(golint ./... | tee /dev/stderr)"
+	GOLINT_FAIL=$?
+	echo
+fi
 
 echo "### go vet"
 go vet ./...
@@ -60,44 +68,57 @@ if [ -x "$(command -v golangci-lint)" ]; then
 fi
 
 RET=0
-if [ ${GOFMT_FAIL} -eq 1 ] ; then
+if [ ${GOFMT_FAIL} -eq 1 ]; then
 	echo "gofmt:         FAIL"
 	RET=1
 else
 	echo "gofmt:         PASS"
 fi
-if [ ${GOIMPORTS_FAIL} -eq 1 ] ; then
+if [ ${GOIMPORTS_FAIL} -eq 1 ]; then
 	echo "goimports:     FAIL"
 	RET=1
 else
 	echo "goimports:     PASS"
 fi
-if [ ${GOLINT_FAIL} -eq 1 ] ; then
-	echo "golint:        FAIL"
-	#RET=1
+if [ -f .revive.toml ]; then
+	if [ ${REVIVE_FAIL} -eq 1 ]; then
+		echo "revive:        FAIL"
+		# We do consider a revive lint failure a CI failure, since we can disable
+		# specific rules using the .revive.toml config.
+		RET=1
+	else
+		echo "revive:        PASS"
+	fi
 else
-	echo "golint:        PASS"
+	if [ ${GOLINT_FAIL} -eq 1 ]; then
+		echo "golint:        FAIL"
+		# We don't consider a golint failure a CI failure.
+		#RET=1
+	else
+		echo "golint:        PASS"
+	fi
 fi
-if [ ${GOVET_FAIL} -eq 1 ] ; then
+if [ ${GOVET_FAIL} -eq 1 ]; then
 	echo "go vet:        FAIL"
 	RET=1
 else
 	echo "go vet:        PASS"
 fi
-if [ ${GOTEST_FAIL} -eq 1 ] ; then
+if [ ${GOTEST_FAIL} -eq 1 ]; then
 	echo "go test:       FAIL"
 	RET=1
 else
 	echo "go test:       PASS"
 fi
-if [ ${GOTEST_RACE_FAIL} -eq 1 ] ; then
+if [ ${GOTEST_RACE_FAIL} -eq 1 ]; then
 	echo "go test -race: FAIL"
 	RET=1
 else
 	echo "go test -race: PASS"
 fi
-if [ ${GOLANGCI_LINT_FAIL} -eq 1 ] ; then
+if [ ${GOLANGCI_LINT_FAIL} -eq 1 ]; then
 	echo "GolangCI-Lint: FAIL"
+	# We don't consider a Golang-CI lint failure a CI failure.
 	#RET=1
 else
 	echo "GolangCI-Lint: PASS"
